Added adbp flag and fixed bugs
