+++ som
@@ -72,6 +72,8 @@ tools.add_argument('-wd', '--web-debugger', help='start Web debugger',
                     dest='web_debugger', action='store_true', default=False)
 tools.add_argument('-ad', '--assisted-debugging', help='enable assisted debugging',
                     dest='assisted_debugging', action='store_true', default=False)
+tools.add_argument('-adbp', '--assisted-debugging-breakpoints', help='set assisted debugging breakpoints limit',
+                    dest='assisted_debugging_breakpoints', default=0)
 tools.add_argument('-dm', '--dynamic-metrics', help='Capture Dynamic Metrics',
                     dest='dynamic_metrics', action='store_true', default=False)
 tools.add_argument('-si', '--si-candidates', help='Identify candidates for super-instructions',
@@ -311,7 +313,7 @@ if args.coverage:
 if args.java_coverage:
     flags += ['-javaagent:' + BASE_DIR + '/libs/jacoco/lib/jacocoagent.jar=inclbootstrapclasses=true,destfile=' + args.java_coverage]
 
-if args.web_debugger or args.assisted_debugging:
+if args.web_debugger:
     SOM_ARGS += ['--web-debug']
 if args.dynamic_metrics:
     SOM_ARGS += ['--dynamic-metrics']
@@ -351,14 +353,19 @@ if args.actor_snapshots_all:
     flags += ['-Dsom.actorSnapshotAll=true']
 
 if (args.truffle_profile or args.web_debugger or
-    args.dynamic_metrics or args.coverage or args.si_candidates or args.assisted_debugging):
+    args.dynamic_metrics or args.coverage or args.si_candidates):
     flags += ['-Dsom.instrumentation=true']
 
-if args.web_debugger or args.assisted_debugging:
+if args.web_debugger:
     flags += ['-Dsom.truffleDebugger=true']
 
 if args.assisted_debugging:
     flags += ['-Dsom.assistedDebugging=true']
+
+if args.assisted_debugging_breakpoints:
+    flags += ['-Dsom.assistedDebugging=true']
+if args.assisted_debugging_breakpoints:
+    flags += ['-Dsom.assistedDebuggingBp=%s' % args.assisted_debugging_breakpoints ]
 
 if not args.interpreter and args.perf_warnings:
     flags += ['-Dgraal.TruffleCompilationExceptionsAreFatal=true',
